refactor: 拡張関数などをファイルに分割 (#4)
