+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/util/AndroidViewModelExtentions.kt
@@ -1,0 +1,7 @@
+package jp.co.yumemi.android.code_check.util
+
+import android.content.Context
+import androidx.lifecycle.AndroidViewModel
+
+val AndroidViewModel.context: Context
+    get() = getApplication()
