+++ app/src/main/res/navigation/nav_graph.xml
@@ -22,7 +22,7 @@
         tools:layout="@layout/fragment_detail">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.Repository" />
+            app:argType="jp.co.yumemi.android.code_check.models.Repository" />
     </fragment>
 
 </navigation>
